requirements.txt
cbor2==5.3.0
Deprecated==1.2.12
luma.core==2.3.1
luma.led-matrix==1.5.0
Pillow==8.2.0
pkg-resources==0.0.0
pyftdi==0.53.1
pyserial==3.5
pyusb==1.1.1
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
smbus2==0.4.1
spidev==3.5
wrapt==1.12.1
ws2812==1.0.0
